search: split functions into separate or related files
Updates #82.

Also update copyright dates.
